// src/types/closure.ts
export interface ClosureRecord {
  _id: number;
  name: string;
  address_myenv: string;
  description_myenv: string;
  photourl?: string;
  q1_cleaningstartdate: string;
  q1_cleaningenddate: string;
  q2_cleaningstartdate: string;
  q2_cleaningenddate: string;
  q3_cleaningstartdate: string;
  q3_cleaningenddate: string;
  q4_cleaningstartdate: string;
  q4_cleaningenddate: string;
  other_works_startdate: string;
  other_works_enddate: string;
}